Refactor view queries to inherit shared code
